+++ xmake/core/sandbox/modules/coroutine.lua
@@ -45,7 +45,7 @@ function sandbox_coroutine.resume(co, ...)
 
         -- get errors
         local errors = results
-        if option.get("backtrace") then
+        if option.get("diagnosis") then
             errors = debug.traceback(co, results)
         elseif type(results) == "string" then
             -- remove the prefix info