+++ xmake/core/sandbox/modules/try.lua
@@ -35,7 +35,7 @@ local sandbox_try = sandbox_try or {}
 function sandbox_try._traceback(errors)
 
     -- not verbose?
-    if not option.get("backtrace") then
+    if not option.get("diagnosis") then
         if errors then
             -- remove the prefix info
             local _, pos = errors:find(":%d+: ")