+++ xmake/core/sandbox/modules/utils.lua
@@ -143,14 +143,14 @@ function sandbox_utils.cprintf(format, ...)
     end
 end
 
--- print() if enable verbose
+-- print the verbose information 
 function sandbox_utils.vprint(format, ...)
     if option.get("verbose") then
         sandbox_utils.print(format, ...)
     end
 end
 
--- vprintf() if enable verbose
+-- print the verbose information without newline
 function sandbox_utils.vprintf(format, ...)
     if option.get("verbose") then
         sandbox_utils.printf(format, ...)