+++ CHANGELOG.md
@@ -13,6 +13,7 @@
 * Improve compilation dependences
 * Support *.xz for extractor
 * [#249](https://github.com/tboox/xmake/pull/249): revise progress formatting to space-leading three digit percentages 
+* [#247](https://github.com/tboox/xmake/pull/247): Add `-D` and `--diagnosis` instead of `--backtrace`
 
 ## v2.2.2
 
@@ -511,6 +512,7 @@
 * 改进编译依赖，对源文件列表的改动进行依赖判断
 * 支持解压*.xz文件
 * [#249](https://github.com/tboox/xmake/pull/249): 改进编译进度信息显示格式
+* [#247](https://github.com/tboox/xmake/pull/247): 添加`-D`和`--diagnosis`去替换`--backtrace`，改进诊断信息显示
 
 ## v2.2.2